requirements.txt
pdf2image==1.16.0
opencv-python==4.7.0.72
pytesseract==0.3.10
matplotlib>=3.5.3
transformers==4.26.1
torch>=1.9.0
pypdf2==3.0.1
scikit-learn==1.0.2
pycryptodome==3.18.0
pdfminer==20191125
pybind11==2.10.4
poppler-utils==0.1.0
python-poppler==0.3.0
